Converted to SASS, added About page
