+++ src/App.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import { BrowserRouter as Router, Routes, Route, } from "react-router-dom";
 import Home from './pages/Home';
+import About from './pages/About';
 import Twitch from './pages/Twitch';
 import FourOhFour from './pages/FourOhFour';
 
@@ -11,6 +12,7 @@ const App = () => {
         <Routes>
           <Route index element={<Home />} />
           <Route exact path="/" element={<Home />} />
+          <Route exact path="/about" element={<About />} />
           <Route path="/twitch" element={<Twitch />} />
           <Route path="*" element={<FourOhFour />} />
         </Routes>
