+++ src/_global.scss
@@ -1,0 +1,7 @@
+/* Fonts */
+$global-font-family: 'Outfit';
+
+/* Global colours */
+$default-bg-color: #222;
+$navbar-bg-color: #161513;
+$footer-bg-color: #161513;
