+++ src/components/Navbar/Navbar.scss
@@ -1,8 +1,18 @@
+@use '../../global' as *;
+
+header.main {
+    background-color: $navbar-bg-color;
+}
+
 nav.global {
     display: flex;
     align-items: center;
     justify-content: space-between;
-    margin: 20px 170px;
+    margin: 0 170px;
+
+    img.header-logo {
+        height: 50px;
+    }
 }
 
 .nav-menu {
@@ -11,18 +21,19 @@ nav.global {
     list-style: none;
     gap: 50px;
     font-size: 20px;
-}
 
-.nav-menu li {
-    display: flex;
-    flex-direction: column;
-    gap: 5px;
-    cursor: pointer;
-}
+    li {
+        display: flex;
+        flex-direction: column;
+        gap: 5px;
+        cursor: pointer;
 
-.nav-menu li img {
-    display: flex;
-    margin: auto;
+        img {
+            display: flex;
+            margin: auto;
+        }
+    }
+
 }
 
 .anchor-link {
