+++ src/index.scss
@@ -1,0 +1,13 @@
+@use 'global' as *;
+
+* {
+  margin: 0;
+  padding: 0;
+}
+
+body {
+  height: 100vh;
+  background: $default-bg-color;
+  color: white;
+  font-family: $global-font-family, sans-serif;
+}
