+++ src/pages/About.scss
@@ -1,3 +1,5 @@
+@use '../global' as *;
+
 .about {
     display: flex;
     flex-direction: column;
@@ -49,7 +51,6 @@
     gap: 20px;
 }
 
-/* Individual skills */
 .about-skill {
     display: flex;
     gap: 50px;
@@ -84,7 +85,6 @@
     margin-bottom: 80px;
 }
 
-/* Individual achivements */
 .about-achievement {
     display: flex;
     flex-direction: column;
