+++ src/pages/Home.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Navbar from '../components/Navbar/Navbar';
 import Hero from '../components/Hero/Hero';
-import About from '../components/About/About';
+// import About from '../components/About/About';
 import Services from '../components/Services/Services';
 import MyWork from '../components/MyWork/MyWork';
 import Contact from '../components/Contact/Contact';
@@ -15,7 +15,6 @@ const Home = () => {
       </header>
       <main>
         <Hero />
-        <About />
         <Services />
         <MyWork />
         <Contact />
